Refactor flags
